fix(bytes): add meta index sequence and safer handling
